requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.3
Werkzeug==0.9.4
feedparser==5.1.3
gevent==1.0
greenlet==0.4.2
ipython==1.2.1
itsdangerous==0.23
path.py==5.1
